feat: support for new architecture
