+++ Example/testHotUpdate/.ruby-version
@@ -1,1 +1,1 @@
-2.7.6
+3.1.1
