+++ Example/testHotUpdate/android/app/build.gradle
