+++ Example/testHotUpdate/android/build.gradle
