+++ Example/testHotUpdate/ios/.xcode.env
@@ -2,6 +2,7 @@
 # used when running script phases inside Xcode.
 # To customize your local environment, you can create an `.xcode.env.local`
 # file that is not versioned.
+
 # NODE_BINARY variable contains the PATH to the node executable.
 #
 # Customize the NODE_BINARY variable here.
