+++ Example/testHotUpdate/ios/testHotUpdate/AppDelegate.mm
