+++ ios/RCTPushy/RCTPushy.m
@@ -1,7 +1,10 @@
 #import "RCTPushy.h"
 #import "RCTPushyDownloader.h"
 #import "RCTPushyManager.h"
-
+// Thanks to this guard, we won't import this header when we build for the old architecture.
+#ifdef RCT_NEW_ARCH_ENABLED
+#import "RNUpdateSpec.h"
+#endif
 
 #import <React/RCTConvert.h>
 #import <React/RCTLog.h>
@@ -561,5 +564,14 @@ RCT_EXPORT_METHOD(markSuccess)
     return buildTime;
 #endif
 }
+
+// Thanks to this guard, we won't compile this code when we build for the old architecture.
+#ifdef RCT_NEW_ARCH_ENABLED
+- (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:
+    (const facebook::react::ObjCTurboModule::InitParams &)params
+{
+    return std::make_shared<facebook::react::NativeCalculatorSpecJSI>(params);
+}
+#endif
 
 @end
