+++ lib/main.js
@@ -15,25 +15,27 @@ const {
   version: v,
 } = require('react-native/Libraries/Core/ReactNativeVersion');
 const RNVersion = `${v.major}.${v.minor}.${v.patch}`;
-
-let Pushy = NativeModules.Pushy;
+const isTurboModuleEnabled = global.__turboModuleProxy != null;
+const Pushy = isTurboModuleEnabled ?
+  require("./NativeUpdate").default :
+  NativeModules.Pushy;
 
 if (!Pushy) {
   throw new Error('react-native-update模块无法加载，请对照安装文档检查配置。');
 }
 
-export const downloadRootDir = Pushy.downloadRootDir;
-export const packageVersion = Pushy.packageVersion;
-export const currentVersion = Pushy.currentVersion;
-export const isFirstTime = Pushy.isFirstTime;
-const rolledBackVersion = Pushy.rolledBackVersion;
+export const downloadRootDir = isTurboModuleEnabled ? Pushy.getConstants().downloadRootDir : Pushy.downloadRootDir;
+export const packageVersion = isTurboModuleEnabled ? Pushy.getConstants().packageVersion : Pushy.packageVersion;
+export const currentVersion = isTurboModuleEnabled ? Pushy.getConstants().currentVersion : Pushy.currentVersion;
+export const isFirstTime = isTurboModuleEnabled ? Pushy.getConstants().isFirstTime : Pushy.isFirstTime;
+const rolledBackVersion = isTurboModuleEnabled ? Pushy.getConstants().rolledBackVersion : Pushy.rolledBackVersion;
 export const isRolledBack = typeof rolledBackVersion === 'string';
 
-export const buildTime = Pushy.buildTime;
-let blockUpdate = Pushy.blockUpdate;
-let uuid = Pushy.uuid;
+export const buildTime = isTurboModuleEnabled ? Pushy.getConstants().buildTime : Pushy.buildTime;
+let blockUpdate = isTurboModuleEnabled ? Pushy.getConstants().blockUpdate : Pushy.blockUpdate;
+let uuid = isTurboModuleEnabled ? Pushy.getConstants().uuid : Pushy.uuid;
 
-if (Platform.OS === 'android' && !Pushy.isUsingBundleUrl) {
+if (Platform.OS === 'android' && !(isTurboModuleEnabled ? Pushy.getConstants().isUsingBundleUrl : Pushy.isUsingBundleUrl)) {
   throw new Error(
     'react-native-update模块无法加载，请对照文档检查Bundle URL的配置',
   );
