+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-update",
-  "version": "8.1.0",
+  "version": "9.1.0",
   "description": "react-native hot update",
   "main": "lib/index.js",
   "scripts": {
@@ -30,5 +30,14 @@
   "homepage": "https://github.com/reactnativecn/react-native-pushy#readme",
   "dependencies": {
     "nanoid": "^3.3.3"
-  }
+  },
+  "codegenConfig": {
+    "libraries": [
+        {
+        "name": "RNUpdateSpec",
+        "type": "modules",
+        "jsSrcsDir": "lib"
+        }
+    ]
+}
 }
